requirements.txt
ultralytics
streamlit
opencv-python
pillow
pandas
numpy
matplotlib
fpdf2
torch
torchvision
pyyaml
